fix: 헤더바 모달창 구현된 파일 CircleMap -> Circle
