+++ components/header/CircleHeader.jsx
@@ -1,14 +1,46 @@
-import { Text, View, StyleSheet, TouchableOpacity } from 'react-native';
+import { Text, View, TouchableOpacity } from 'react-native';
 import { FontAwesome, Ionicons } from '@expo/vector-icons';
 import { styles } from './styles';
 import { useRecoilState } from 'recoil';
+import { useState, useMemo } from 'react';
 import { userState } from '../../stores/user-store';
 import { Image } from 'expo-image';
 import { useNavigation } from '@react-navigation/native';
+import { BottomModal } from '../Modal/Modal';
 
 export const CircleHeader = () => {
   const [user, setUser] = useRecoilState(userState);
+  const [isModalVisible, setModalVisible] = useState(false);
+  const toggleModal = () => {
+    setModalVisible(!isModalVisible);
+    console.log("동작됨");
+  };
   const navigation = useNavigation();
+  const photoOptions = useMemo(
+    () => [
+      {
+        text: '유저 추가',
+        icon: require('../../assets/icons/add_user_icon.png'),
+        iconStyle: styles.user_add,
+        textStyle: {},
+        // onPress:
+      },
+      {
+        text: '유저 정렬',
+        icon: require('../../assets/icons/filter_user_icon.png'),
+        iconStyle: styles.user_array,
+        textStyle: {},
+        // onPress
+      },
+      {
+        text: '써클 이름 변경',
+        icon: require('../../assets/icons/edit_circle_name.png'),
+        iconStyle: styles.circle_name,
+        textStyle: {}
+        // onPress
+      }
+    ]
+  );
   return (
     <View style={styles.container}>
       <Text style={{ fontSize: 20, fontFamily: 'IropkeBatang', color: '#44403C' }}>PicMap</Text>
@@ -26,7 +58,8 @@ export const CircleHeader = () => {
             <FontAwesome name="user-circle-o" style={{ marginLeft: 2 }} size={24} color={'#44403C'} />
           )}
         </TouchableOpacity>
-        <TouchableOpacity>
+        <BottomModal isModalVisible={isModalVisible} toggleModal={toggleModal} buttons={photoOptions} />
+        <TouchableOpacity onPress={toggleModal}>
           <Image source={require('../../assets/icons/circle_array_btn.png')} style={styles.rightHeader} />
         </TouchableOpacity>
       </View>
